requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
coremltools==9.0
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==6.33.5
psutil==7.2.2
pyDeprecate==0.5.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
ruff==0.15.2
scipy==1.17.1
setuptools==82.0.0
six==1.17.0
supervision==0.27.0.post1
sympy==1.14.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
